Breakout external_fixtures
