+++ tests/testing_support/external_fixtures.py
@@ -13,128 +13,7 @@
 # limitations under the License.
 
 from newrelic.api.external_trace import ExternalTrace
-from newrelic.api.transaction import current_transaction
-from newrelic.common.encoding_utils import (json_encode, json_decode,
-    obfuscate, deobfuscate, DistributedTracePayload)
-from newrelic.common.object_wrapper import (transient_function_wrapper,
-        function_wrapper)
-
-OUTBOUND_TRACE_KEYS_REQUIRED = (
-        'ty', 'ac', 'ap', 'tr', 'pr', 'sa', 'ti')
-
-
-def validate_outbound_headers(header_id='X-NewRelic-ID',
-        header_transaction='X-NewRelic-Transaction'):
-    transaction = current_transaction()
-    headers = transaction._test_request_headers
-    settings = transaction.settings
-    encoding_key = settings.encoding_key
-
-    assert header_id in headers
-
-    values = headers[header_id]
-    if isinstance(values, list):
-        assert len(values) == 1, headers
-        assert isinstance(values[0], type(''))
-        value = values[0]
-    else:
-        value = values
-
-    cross_process_id = deobfuscate(value, encoding_key)
-    assert cross_process_id == settings.cross_process_id
-
-    assert header_transaction in headers
-
-    values = headers[header_transaction]
-    if isinstance(values, list):
-        assert len(values) == 1, headers
-        assert isinstance(values[0], type(''))
-        value = values[0]
-    else:
-        value = values
-
-    (guid, record_tt, trip_id, path_hash) = \
-            json_decode(deobfuscate(value, encoding_key))
-
-    assert guid == transaction.guid
-    assert record_tt == transaction.record_tt
-    assert trip_id == transaction.trip_id
-    assert path_hash == transaction.path_hash
-
-
-def validate_distributed_tracing_header(header='newrelic'):
-    transaction = current_transaction()
-    headers = transaction._test_request_headers
-    account_id = transaction.settings.account_id
-    application_id = transaction.settings.primary_application_id
-
-    assert header in headers, headers
-
-    values = headers[header]
-    if isinstance(values, list):
-        assert len(values) == 1, headers
-        assert isinstance(values[0], type(''))
-        value = values[0]
-    else:
-        value = values
-
-    # Parse payload
-    payload = DistributedTracePayload.from_http_safe(value)
-
-    # Distributed Tracing v0.1 is currently implemented
-    assert payload['v'] == [0, 1], payload['v']
-
-    data = payload['d']
-
-    # Verify all required keys are present
-    assert all(k in data for k in OUTBOUND_TRACE_KEYS_REQUIRED)
-
-    # Type will always be App (not mobile / browser)
-    assert data['ty'] == 'App'
-
-    # Verify account/app id
-    assert data['ac'] == account_id
-    assert data['ap'] == application_id
-
-    # Verify data belonging to this transaction
-    assert data['tx'] == transaction.guid
-
-    # If span events are enabled, id should be sent
-    # otherwise, id should be omitted
-    if transaction.settings.span_events.enabled:
-        assert 'id' in data
-    else:
-        assert 'id' not in data
-
-    # Verify referring transaction information
-    assert len(data['tr']) == 32
-    if transaction.referring_transaction_guid is not None:
-        assert data['tr'] == transaction._trace_id
-    else:
-        assert data['tr'].startswith(transaction.guid)
-
-    assert 'pa' not in data
-
-    # Verify timestamp is an integer
-    assert isinstance(data['ti'], int)
-
-    # Verify that priority is a float
-    assert isinstance(data['pr'], float)
-
-
-@function_wrapper
-def validate_cross_process_headers(wrapped, instance, args, kwargs):
-    result = wrapped(*args, **kwargs)
-
-    transaction = current_transaction()
-    settings = transaction.settings
-
-    if settings.distributed_tracing.enabled:
-        validate_distributed_tracing_header()
-    else:
-        validate_outbound_headers()
-
-    return result
+from newrelic.common.object_wrapper import transient_function_wrapper
 
 
 def create_incoming_headers(transaction):
@@ -163,36 +42,6 @@ def create_incoming_headers(transaction):
 
     return headers
 
-
-def validate_external_node_params(params=[], forgone_params=[]):
-    """
-    Validate the parameters on the external node.
-
-    params: a list of tuples
-    forgone_params: a flat list
-    """
-    @transient_function_wrapper('newrelic.api.external_trace',
-            'ExternalTrace.process_response_headers')
-    def _validate_external_node_params(wrapped, instance, args, kwargs):
-        result = wrapped(*args, **kwargs)
-
-        # This is only validating that logic to extract cross process
-        # header and update params in ExternalTrace is succeeding. This
-        # is actually done after the ExternalTrace __exit__() is called
-        # with the ExternalNode only being updated by virtue of the
-        # original params dictionary being aliased rather than copied.
-        # So isn't strictly validating that params ended up in the actual
-        # ExternalNode in the transaction trace.
-
-        for name, value in params:
-            assert instance.params[name] == value
-
-        for name in forgone_params:
-            assert name not in instance.params
-
-        return result
-
-    return _validate_external_node_params
 
 def validate_synthetics_external_trace_header(required_header=(),
         should_exist=True):
@@ -242,101 +91,3 @@ def validate_synthetics_external_trace_header(required_header=(),
         return result
 
     return _validate_synthetics_external_trace_header
-
-def validate_outbound_headers(header_id='X-NewRelic-ID',
-        header_transaction='X-NewRelic-Transaction'):
-    transaction = current_transaction()
-    headers = transaction._test_request_headers
-    settings = transaction.settings
-    encoding_key = settings.encoding_key
-
-    assert header_id in headers
-
-    values = headers[header_id]
-    if isinstance(values, list):
-        assert len(values) == 1, headers
-        assert isinstance(values[0], type(''))
-        value = values[0]
-    else:
-        value = values
-
-    cross_process_id = deobfuscate(value, encoding_key)
-    assert cross_process_id == settings.cross_process_id
-
-    assert header_transaction in headers
-
-    values = headers[header_transaction]
-    if isinstance(values, list):
-        assert len(values) == 1, headers
-        assert isinstance(values[0], type(''))
-        value = values[0]
-    else:
-        value = values
-
-    (guid, record_tt, trip_id, path_hash) = \
-            json_decode(deobfuscate(value, encoding_key))
-
-    assert guid == transaction.guid
-    assert record_tt == transaction.record_tt
-    assert trip_id == transaction.trip_id
-    assert path_hash == transaction.path_hash
-
-
-def validate_distributed_tracing_header(header='newrelic'):
-    transaction = current_transaction()
-    headers = transaction._test_request_headers
-    account_id = transaction.settings.account_id
-    application_id = transaction.settings.primary_application_id
-
-    assert header in headers, headers
-
-    values = headers[header]
-    if isinstance(values, list):
-        assert len(values) == 1, headers
-        assert isinstance(values[0], type(''))
-        value = values[0]
-    else:
-        value = values
-
-    # Parse payload
-    payload = DistributedTracePayload.from_http_safe(value)
-
-    # Distributed Tracing v0.1 is currently implemented
-    assert payload['v'] == [0, 1], payload['v']
-
-    data = payload['d']
-
-    # Verify all required keys are present
-    assert all(k in data for k in OUTBOUND_TRACE_KEYS_REQUIRED)
-
-    # Type will always be App (not mobile / browser)
-    assert data['ty'] == 'App'
-
-    # Verify account/app id
-    assert data['ac'] == account_id
-    assert data['ap'] == application_id
-
-    # Verify data belonging to this transaction
-    assert data['tx'] == transaction.guid
-
-    # If span events are enabled, id should be sent
-    # otherwise, id should be omitted
-    if transaction.settings.span_events.enabled:
-        assert 'id' in data
-    else:
-        assert 'id' not in data
-
-    # Verify referring transaction information
-    assert len(data['tr']) == 32
-    if transaction.referring_transaction_guid is not None:
-        assert data['tr'] == transaction._trace_id
-    else:
-        assert data['tr'].startswith(transaction.guid)
-
-    assert 'pa' not in data
-
-    # Verify timestamp is an integer
-    assert isinstance(data['ti'], int)
-
-    # Verify that priority is a float
-    assert isinstance(data['pr'], float)
